✨ feat(闲置检测): 现在设备会在长时间闲置后自动执行一些事件
现在在闲置一段时间后，会根据注册列表执行特定事件，检测到任何摇杆、按键、CDC信号输入都会使得设备退出闲置状态。在默认配置下，闲置10分钟后会自动关闭所有灯光，在退出闲置状态时会播放设备启动自检灯光。

+++ src/lib/timeout.c
@@ -5,15 +5,15 @@
 #define MAX_TIMER_COUNT  16
 
 typedef struct {
-  uint16_t time;
-  uint16_t period;
+  uint32_t time;
+  uint32_t period;
   void (*callback)(void);
 } Timer_TypeDef;
 
 Timer_TypeDef interval[MAX_TIMER_COUNT] = {0};
 Timer_TypeDef timeout[MAX_TIMER_COUNT]  = {0};
 
-uint16_t timerSet = 0;
+uint32_t timerSet = 0;
 
 xdata void Timer4_Config(void)
 {
@@ -49,7 +49,7 @@ void TIM4_IRQHandler(void)
   }
 }
 
-uint8_t setInterval(void (*callback)(void), uint16_t period)
+uint8_t setInterval(void (*callback)(void), uint32_t period)
 {
   for (uint8_t i = 0; i < MAX_TIMER_COUNT; i++) {
     if (interval[i].callback == NULL) {
@@ -62,7 +62,7 @@ uint8_t setInterval(void (*callback)(void), uint16_t period)
   return 0xFF;
 }
 
-uint8_t setTimeout(void (*callback)(void), uint16_t period)
+uint8_t setTimeout(void (*callback)(void), uint32_t period)
 {
   for (uint8_t i = 0; i < MAX_TIMER_COUNT; i++) {
     if (timeout[i].callback == NULL) {
