+++ src/lib/timeout.h
@@ -5,10 +5,10 @@
 
 void Timeout_Init();
 
-uint8_t setInterval(void (*callback)(void), uint16_t period);
+uint8_t setInterval(void (*callback)(void), uint32_t period);
 void clearInterval(uint8_t id);
 
-uint8_t setTimeout(void (*callback)(void), uint16_t period);
+uint8_t setTimeout(void (*callback)(void), uint32_t period);
 void clearTimeout(uint8_t id);
 
 void Timer_Process();
